+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Entity-entity pushback is no longer applied in forced movement states like rolling and leaping.
 
 ### Removed